+++ tests/BigCommerce/Api/Widgets/WidgetTemplateApiTest.php
@@ -9,9 +9,12 @@ class WidgetTemplateApiTest extends BigCommerceApiTest
     public function testCanGetTemplate(): void
     {
         $this->setReturnData('content__widget_templates__get.json');
+        $id = 'c48b131a-ae9d-4767-b5d6-63d9e43bcf75';
 
-        $template = $this->getApi()->widgets()->template('c48b131a-ae9d-4767-b5d6-63d9e43bcf75')->get()->getTemplate();
+        $template = $this->getApi()->widgets()->template($id)->get()->getTemplate();
         $this->assertEquals('Header Images', $template->name);
+        $this->assertEquals("content/widget-templates/$id", $this->getLastRequestPath());
+
     }
 
     public function testCanGetTemplates(): void
@@ -21,5 +24,6 @@ class WidgetTemplateApiTest extends BigCommerceApiTest
         $templates = $this->getApi()->widgets()->templates()->getAll()->getTemplates();
         $this->assertCount(12, $templates);
         $this->assertEquals('{{{html}}}', $templates[2]->template);
+        $this->assertEquals("content/widget-templates", $this->getLastRequestPath());
     }
 }