+++ RELEASE_NOTES.md
@@ -3,3 +3,7 @@
 - Implement the missing _create_ and _update_ endpoints for Brands.
 
 
+### Fixed Issues
+
+- Fixed issue where Widget Template endpoints were incorrect
+